# faster elements(::PermGroup)