Side-effect of the writing-docs project;

```julia
function generate(N::T) where T
    G = PermGroup(N)
    k = zero(T)
    for p in elements(G)
        k += p[1]
    end
    return k
end
```
* old:
  ```julia
  julia> @time generate(10);
    3.818291 seconds (36.29 M allocations: 1.352 GiB, 5.59% gc time)
  
  julia> @time generate(UInt(10));
    3.478949 seconds (36.29 M allocations: 1.352 GiB, 6.14% gc time)
  ```
* new:
  ```julia
  julia> generate(10)
    0.187002 seconds (10.89 M allocations: 332.227 MiB, 24.28% gc time)
  
  julia> generate(UInt(10))
    0.165168 seconds (10.89 M allocations: 332.227 MiB, 24.04% gc time)
  ```

I don't know why, but if nextperm returns just an array (via `return elts, (elts, c, counter+1)`) this cuts the `@time` further to `0.124588 seconds (7.26 M allocations: 221.485 MiB, 26.26% gc time)`. Where are the additional allocations comming from? Why is constructing `perm(elts, false)` so expensive, when for standalone `d = randperm(10)` `perm(d, false)` takes only `15ns`?