@wbhart 2018-05-09
Sometimes the gc needs some extra information to keep track of what's on
the stack (since there may be pointers there it needs to trace). To do this
it makes a small allocation to keep track.

That's usually the reason for this sort of issue. If you can return
something that doesn't contain pointers, it can sometimes optimise better
in my experience.