@kalmarek 2018-05-10
This was way pre-mature; I have a better solution right now, I'll open a new pull